note validators

+++ Sobes/Command/NoteCommand.cs
@@ -1,4 +1,5 @@
 ﻿using DatabaseModels;
+using FluentValidation;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Sobes.Service;
@@ -23,6 +24,12 @@ namespace Sobes.Command
             return await _noteService.GetAllNotesAsync();
         }
     }
+
+    public class GetAllNotesQueryValidator : AbstractValidator<GetAllNotesQuery>
+    {
+        // No validation rules here
+    }
+
     public record CreateNoteCommand(string Title, string Text, List<int> TagIds) : IRequest<int>;
 
     public class CreateNoteCommandHandler : IRequestHandler<CreateNoteCommand, int>
@@ -38,12 +45,36 @@ namespace Sobes.Command
 
         public async Task<int> Handle(CreateNoteCommand request, CancellationToken cancellationToken)
         {
+            var val = new CreateNoteCommandValidator();
+            var res = await val.ValidateAsync(request);
+            if (!res.IsValid)
+            {
+                throw new ArgumentException(res.ToString());
+            }
             var tags = await _context.Tags.Where(t => request.TagIds.Contains(Convert.ToInt32(t.Id))).ToListAsync();
             var note = new Note { Title = request.Title, Text = request.Text, Tags = tags, CreatedDate = DateTime.UtcNow.Date, ModifiedDate = DateTime.UtcNow.Date };
             await _noteService.CreateNoteAsync(note);
             return note.Id;
         }
     }
+
+    public class CreateNoteCommandValidator : AbstractValidator<CreateNoteCommand>
+    {
+        public CreateNoteCommandValidator()
+        {
+            RuleFor(x => x.Title)
+                .NotEmpty().WithMessage("Заголовок не должен быть пустым.")
+                .MaximumLength(255).WithMessage("Длина заголовка не должна превышать 255 символов.");
+
+            RuleFor(x => x.Text)
+                .MaximumLength(255).WithMessage("Длина текста не должна превышать 255 символов.");
+
+            RuleFor(x => x.TagIds)
+                .NotEmpty().WithMessage("Должен быть хотя бы один тэг.")
+                .Must(tagIds => tagIds.All(tagId => tagId > 0)).WithMessage("Все тэги должны иметь допустимые значения.");
+        }
+    }
+
 
     public record DeleteNoteCommand(int Id) : IRequest;
 
@@ -58,9 +89,24 @@ namespace Sobes.Command
 
         public async Task Handle(DeleteNoteCommand request, CancellationToken cancellationToken)
         {
+            var val = new DeleteNoteCommandValidator();
+            var res = await val.ValidateAsync(request);
+            if (!res.IsValid)
+            {
+                throw new ArgumentException(res.ToString());
+            }
             await _noteService.DeleteNoteAsync(request.Id);
         }
     }
+    public class DeleteNoteCommandValidator : AbstractValidator<DeleteNoteCommand>
+    {
+        public DeleteNoteCommandValidator()
+        {
+            RuleFor(x => x.Id)
+                .GreaterThan(0).WithMessage("ID заметки должен быть положительным числом.");
+        }
+    }
+
 
     public record UpdateNoteCommand(int Id, string Title, string Text, List<int> TagIds) : IRequest
     {
@@ -82,16 +128,22 @@ namespace Sobes.Command
 
         public async Task Handle(UpdateNoteCommand request, CancellationToken cancellationToken)
         {
+            var val = new UpdateNoteCommandValidator();
+            var res = await val.ValidateAsync(request);
+            if (!res.IsValid)
+            {
+                throw new ArgumentException(res.ToString());
+            }
             var note = _context.Notes.Find(request.Id);
             if (note == null)
             {
                 throw new Exception("Заметка не найдена");
             }
+
             note.Title = request.Title;
             note.Text = request.Text;
             note.ModifiedDate = DateTime.UtcNow.Date;
 
-            // Обновляем список тегов
             note.Tags = await _context.Tags
                 .Where(t => request.TagIds.Contains(t.Id))
                 .ToListAsync();
@@ -100,10 +152,30 @@ namespace Sobes.Command
 
     }
 
+    public class UpdateNoteCommandValidator : AbstractValidator<UpdateNoteCommand>
+    {
+        public UpdateNoteCommandValidator()
+        {
+            RuleFor(x => x.Id)
+                .GreaterThan(0).WithMessage("ID заметки должен быть положительным числом.");
 
-    public record GetNote(int Id) : IRequest<Note>;
+            RuleFor(x => x.Title)
+                .NotEmpty().WithMessage("Заголовок не должен быть пустым.")
+                .MaximumLength(255).WithMessage("Длина заголовка не должна превышать 255 символов.");
 
-    public class GetNoteCommandHander : IRequestHandler<GetNote, Note>
+            RuleFor(x => x.Text)
+                .MaximumLength(255).WithMessage("Длина текста не должна превышать 255 символов.");
+
+            RuleFor(x => x.TagIds)
+                .NotEmpty().WithMessage("Должен быть хотя бы один тэг.")
+                .Must(tagIds => tagIds.All(tagId => tagId > 0)).WithMessage("Все тэги должны иметь допустимые значения.");
+        }
+    }
+
+
+    public record GetNoteCommand(int Id) : IRequest<Note>;
+
+    public class GetNoteCommandHander : IRequestHandler<GetNoteCommand, Note>
     {
         private readonly NoteService _noteService;
 
@@ -112,11 +184,26 @@ namespace Sobes.Command
             _noteService = noteService;
         }
 
-        public async Task<Note?> Handle(GetNote request, CancellationToken cancellationToken)
+        public async Task<Note?> Handle(GetNoteCommand request, CancellationToken cancellationToken)
         {
-
+            var val = new GetNoteValidator();
+            var res = await val.ValidateAsync(request);
+            if (!res.IsValid)
+            {
+                throw new ArgumentException(res.ToString());
+            }
             return await _noteService.GetNoteAsync(request.Id);
 
         }
     }
+
+    public class GetNoteValidator : AbstractValidator<GetNoteCommand>
+    {
+        public GetNoteValidator()
+        {
+            RuleFor(x => x.Id)
+                .GreaterThan(0).WithMessage("ID заметки должен быть положительным числом.");
+        }
+    }
+
 }
